requirements.txt
bcrypt
boto3
flask
flask-cors
flask-socketio
matplotlib
numpy
PyJWT
